requirements.txt
boto3
sklearn
pyyaml
pandas
numpy
pymongo
dnspython
kneed
xgboost
flask